fix in Sales Order-Receivable Voucher Mapper

+++ erpnext/patches/mar_2012/so_rv_mapper_fix.py
@@ -1,7 +1,10 @@
 def execute():
 	import webnotes
-	webnotes.conn.sql("""DELETE FROM `tabTable Mapper Detail`
+	count = webnotes.conn.sql("""SELECT COUNT(*) FROM `tabTable Mapper Detail`
 		WHERE parent='Sales Order-Receivable Voucher'
 		AND from_table='Sales Order Detail'""")
-	from webnotes.modules.module_manager import reload_doc
-	reload_doc('accounts', 'DocType Mapper', 'Sales Order-Receivable Voucher')
+	if count and count[0][0]==2:
+		webnotes.conn.sql("""DELETE FROM `tabTable Mapper Detail`
+			WHERE parent='Sales Order-Receivable Voucher'
+			AND from_table='Sales Order Detail'
+			AND validation_logic='docstatus = 1'""")
